project/requirements.txt
numpy==1.26.3
scikit-learn==1.3.2
Pillow==10.2.0
seaborn==0.13.2
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
pyyaml==6.0.1
python-dotenv==1.0.1
mlflow==2.10.2
tqdm==4.66.1
pytest==7.4.4
pytest-asyncio==0.23.5
httpx==0.26.0

ipykernel
kagglehub>=0.4.1